Change floats to doubles

+++ ros1/src/data/march_realsense_reader/include/pointcloud_processor/preprocessor.h
@@ -100,7 +100,7 @@ protected:
 
     // Downsampling parameters
     bool voxel_grid_filter;
-    float leaf_size;
+    double leaf_size;
     bool random_filter;
     int remaining_points;
